requirements.txt
PyYAML
transformers
torch
tensorboard
torchvision
datasets
pytest
setproctitle
gradio==4.41.0
evaluate
accelerate
scikit-learn